save a little bit of space
svn path=/trunk/KDE/kdelibs/; revision=632238

+++ kjs/date_object.h
@@ -77,7 +77,7 @@ namespace KJS {
             // non-normative properties (Appendix B)
             GetYear, SetYear, ToGMTString };
     private:
-        int id;
+        short id;
         bool utc;
     };
